Updated install script and added jshint file

+++ install.sh
@@ -3,14 +3,16 @@
 # variables
 dir=~/dotfiles
 olddir=~/dotfiles_old
-files="bash_profile"
+files=("bash_profile" "jscsrc")
 
 # create and change to olddir
 mkdir $olddir
 cd $dir
 
 # move any existing dotfiles to oldir, then symlink to the new ones
-for file in $files; do
-	mv ~/.$file $olddir/$file
-	ln -s $dir/$file ~/.$file
+for file in ${files[*]}; do
+	if [ -f $file ]; then
+		mv ~/.$file $olddir/$file
+		ln -s $dir/$file ~/.$file
+	fi
 done
